+++ src/resolvers/mutation/index.js
@@ -1,4 +1,6 @@
 import authMutation from './auth';
+import closetMutation from './closet';
+import outfitMutation from './outfit';
 import paymentMutation from './payment';
 import requestMutation from './request';
 import subscriptionMutation from './subscription';
@@ -10,6 +12,8 @@ const mutation = {
   ...paymentMutation,
   ...subscriptionMutation,
   ...requestMutation,
+  ...closetMutation,
+  ...outfitMutation,
 };
 
 export default mutation;